requirements.txt
absl-py==2.3.1
attrs==25.4.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.2
cryptography==46.0.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
flatbuffers==25.9.23
fonttools==4.60.1
gcloud==0.18.3
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
httplib2==0.31.0
idna==3.11
jax==0.6.2
jaxlib==0.6.2
jwcrypto==1.5.6
kiwisolver==1.4.9
matplotlib==3.10.7
mediapipe==0.10.9
ml_dtypes==0.5.3
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.3.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==12.0.0
playsound==1.2.2
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pyparsing==3.2.5
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
Pyrebase4==4.8.0
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
python-dateutil==2.9.0.post0
python-jwt==4.1.0
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rsa==4.9.1
scipy==1.15.3
sentencepiece==0.2.1
shiboken6==6.10.0
six==1.17.0
sounddevice==0.5.3
typing_extensions==4.15.0
urllib3==1.26.20
